# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EthereumConnect",
    "description": "NA",
    "version": "1.0"
  },
  "produces": [
    "application/json"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
  "paths": {
    "/ethereum/Status": {
      "get": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "EthereumStatus",
        "tags": [
          "Ethereum Basic"
        ],
        "description": "Check The Status of Ethreum Connected",
        "parameters": [],
        "responses": {}
      }
    },
    "/ethereum/createAccount": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "Account Creation",
        "tags": [
          "Personal"
        ],
        "description": "To create The account",
        "parameters": [
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/unlockAccount": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "Unlock Account",
        "tags": [
          "Personal"
        ],
        "description": "To Unlock The account",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/lockAccount": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "Lock Account",
        "tags": [
          "Personal"
        ],
        "description": "To Lock The account",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/sign": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "sign Transaction",
        "tags": [
          "Personal"
        ],
        "description": "To Lock The account",
        "parameters": [
          {
            "name": "message",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/sendTransaction": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "send Transaction",
        "tags": [
          "Personal"
        ],
        "description": "To send the transaction",
        "parameters": [
          {
            "name": "fromAccount",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "toAccount",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "etherAmount",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/ecRecover": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "recover transaction ",
        "tags": [
          "Personal"
        ],
        "description": "To Lock The account",
        "parameters": [
          {
            "name": "message",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "signature",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/importRawKey": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "importRawKey ",
        "tags": [
          "Personal"
        ],
        "description": "importRawKey",
        "parameters": [
          {
            "name": "priavteKey",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/ethereum/listAccounts": {
      "get": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "List accounts",
        "tags": [
          "Personal"
        ],
        "description": "Get Account Details",
        "parameters": [],
        "responses": {}
      }
    },
    "/miner/setEtherbase": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "For set EtherBase",
        "tags": [
          "Miner"
        ],
        "description": "For set EtherBase",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/Contracts/deployContract": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "for deploy contarct dyanmic ",
        "tags": [
          "Contract"
        ],
        "description": "for deploy contarct dyanmic",
        "content": "multipart/form-data",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          }
              
        ],
        "responses": {}
      }
    },
    "/Contracts/method": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "for transaction ",
        "tags": [
          "Contract"
        ],
        "description": "for deploy contarct dyanmic",
        "content": "multipart/form-data",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          }
              
        ],
        "responses": {}
      }
    },
    "/Contracts/upload": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "for transaction ",
        "tags": [
          "Contract"
        ],
        "description": "for deploy contarct dyanmic",
        "content": "multipart/form-data",
        "parameters": [
          {
            "name": "sampleFile",
            "in": "formData",
            "required": true,
            "type": "file"
          }
              
        ],
        "responses": {}
      }
    },
    "/eth/getTransaction": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "for getting transaction details ",
        "tags": [
          "ETH"
        ],
        "description": "for getting transaction details",
        "content": "multipart/form-data",
        "parameters": [
          {
            "name": "trnsactionHash",
            "in": "formData",
            "required": true,
            "type": "string"
          }
              
        ],
        "responses": {}
      }
    },
    "/eth/getTransactionList": {
      "post": {
        "x-swagger-router-controller": "middleware-name1",
        "operationId": "for getting transaction details ",
        "tags": [
          "ETH"
        ],
        "description": "for getting transaction details",
        "content": "multipart/form-data",
        "parameters": [
          {
            "name": "accountAddress",
            "in": "formData",
            "required": true,
            "type": "string"
          }
              
        ],
        "responses": {}
      }
    }

  }
}